monai_ml/app/requirements.txt
fastapi[all]
python-socketio
fastapi-socketio
uvicorn
pyyaml
sse-starlette
dcm2niix